requirements.txt
pyparallel
pyserial
PyQt5==5.13.2
PyQtGraph
argparse
pyyaml
qdarkstyle
pyfirmata

